Add files via upload
The Iris dataset is a well-known dataset in the field of machine learning and statistics. It contains 150 samples of iris flowers, with 50 samples each from three species: Iris-setosa, Iris-versicolor, and Iris-virginica. Each sample has four features: Sepal Length, Sepal Width, Petal Length, and Petal Width. The goal is to classify the species of the iris flower based on these features. In this project, we used a neural network implemented in PyTorch to classify the iris species.